requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
openai==0.27.8
gtts==2.3.2
python-dotenv==1.0.0
transformers==4.33.3
sentencepiece==0.1.99
PyPDF2==3.0.0
python-docx==0.8.11
python-multipart